CodersCartel_KDSH_2026/code/requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.41.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentence-transformers>=2.7.0
faiss-cpu
tqdm>=4.65.0
